Refactor: rename `ConflictModals` -> `ConflictModal`
